update eloquent relation

+++ app/Http/Controllers/PostLikesController.php
@@ -25,6 +25,14 @@ class PostLikesController extends Controller
             $validated = $request->validate([
                 'post_id' => ['required'],  
             ]);
+            $userHasLike = PostLike::where('user_id',Auth::user()->id)->where('post_id',$request->input('post_id'))->get();
+            if(count($userHasLike) > 0){
+                return response([
+                    'success' => true,
+                    'message' => 'User Had been',
+                    'data' => $userHasLike
+                ], 201);
+            }
             $post = PostLike::create([
                 'user_id' => Auth::user()->id,
                 'post_id' =>$request->input('post_id'),
@@ -44,10 +52,10 @@ class PostLikesController extends Controller
          * @param  int  $id
          * @return \Illuminate\Http\Response
          */
-        public function destroy(PostLike $postLike)
+        public function destroy(Request $request,$post_id)
         {
-            //
-            $postLike->delete();
+            $postLike = PostLike::where('user_id',Auth::user()->id)->where('post_id',$post_id)->delete();
+            // $postLike->delete();
             return response([
                 'success' => true,
                 'message' => 'Post Like Has been deleted',
